Create README.md
Hello, here is my first project I don't know if it's good or right, but I'll be happy for the feedback for some improvements.

Thank you very much for the teaching! 





It didn't have a good basis to replace, but I used resources that were shared in the course; CatsVSDog.